// nodes/CapmonsterCloud/types.ts
export type TaskType =
	| 'json'
	| 'recaptchaV2Enterprise'
	| 'recaptchaV3Enterprise'
	| 'recaptchaV2'
	| 'recaptchaV3'
	| 'imageToText'
	| 'geeTestV3'
	| 'geeTestV4'
	| 'recaptchaClick'
	| 'turnstile'
	| 'turnstileChallengeToken'
	| 'turnstileChallengeCfClearance'
	| 'turnstileWaitRoom'
	| 'complexImageRecognition'
	| 'datadome'
	| 'basilisk'
	| 'tendi'
	| 'amazonJsApi'
	| 'amazonFullChallenge'
	| 'amazonInvisibleChallenge'
	| 'binance'
	| 'imperva'
	| 'prosopo'
	| 'temu'
	| 'yidun'
	| 'mtcaptcha'
	| 'altcha'
	| 'funcaptcha'
	| 'castle'
	| 'tspd'
	| 'hunt';